pyproject.toml
[project]
name = "cce-hack"
version = "0.1.0"
description = "CCE mooring analytics + forecasting (DataHacks / dual-track)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "streamlit>=1.28",
  "streamlit-extras>=0.4.0",
  "anthropic>=0.40.0",
  "marimo>=0.8.0",
  "matplotlib>=3.7",
  "plotly>=5.18",
  "pydeck>=0.8.0",
  "statsmodels>=0.14",
  "PyCO2SYS>=0.15",
  "PyWavelets>=1.4",
  "shap>=0.44",
  "umap-learn>=0.5",
  "hdbscan>=0.8",
]

[project.optional-dependencies]
docs = [
  "sphinx>=7.0",
  "pydata-sphinx-theme>=0.15.0",
]
netcdf = ["xarray>=2023.1", "netCDF4>=1.6"]
deep = [
  "tensorflow-cpu>=2.15",
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
